maintenance fixes

+++ include/bd/tbb/parallelreduce_minmax.h
@@ -50,7 +50,7 @@ public:
       
       if (val<min_value) { min_value = val; }
       if (val>max_value) { max_value = val; }
-      tot_value += val;
+      tot_value += static_cast<double>(val);
     }
   }
